backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.9
pandas==2.0.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
numpy>=1.26.0
langchain==0.1.0
langchain-openai==0.0.7
langchain-community==0.0.10